Fixes components moving when running an image
The components of the secondary window were misplaced after running a
container. This was caused because they were removed and replaced by
the components for the running panel.

This problem has been solved by showing the running panel's
components in an independent window instead of replacing the
components in the secondary window.