Adds support for CLI/GUI execution
The secondary window has been updated to support the execution of the
CLI and GUI commands of the images, when supported.

To do so, this window was completely rebuilt, reviewing its behaviour
to avoid inconsistencies.

The run window has also been updated to make it more interactive and
allow stopping the execution of the command.